use google map replace button and show map

+++ app/src/main/java/dev/leonardpark/app/weatherapp/model/Coordinate.kt
@@ -1,6 +1,6 @@
 package dev.leonardpark.app.weatherapp.model
 
 data class Coordinate(
-  var lon: Float,
-  var lat: Float
+  var lon: Double,
+  var lat: Double
 )
